+++ src/camera.rs
@@ -1,23 +1,23 @@
-use std::f64::consts::{FRAC_PI_3};
+use std::f32::consts::{FRAC_PI_3};
 
 use cgmath::{EuclideanSpace, InnerSpace, Matrix, Matrix4, PerspectiveFov, Point3, Rad, Vector3};
 
 
-pub const NEAR_PLANE:    f64 = 1.0;
-pub const FAR_PLANE:     f64 = 60.0;
-pub const FIELD_OF_VIEW: f64 = FRAC_PI_3;
+pub const NEAR_PLANE:    f32 = 1.0;
+pub const FAR_PLANE:     f32 = 60.0;
+pub const FIELD_OF_VIEW: f32 = FRAC_PI_3;
 
 #[derive(Clone)]
 pub struct Camera {
-	pub view:         Matrix4<f64>,
-	pub projection:   Matrix4<f64>,
-	pub aspect_ratio: f64,
+	pub view:         Matrix4<f32>,
+	pub projection:   Matrix4<f32>,
+	pub aspect_ratio: f32,
 }
 
 impl Camera {
-	pub fn new(center: Point3<f64>, view_direction: Vector3<f64>, aspect_ratio: f64) -> Camera {
-		const DISTANCE: f64 = 10.0;
-		const CENTER_OFFSET: Vector3<f64> = Vector3{ x: 0.0, y: 5.0, z: 0.0};
+	pub fn new(center: Point3<f32>, view_direction: Vector3<f32>, aspect_ratio: f32) -> Camera {
+		const DISTANCE: f32 = 10.0;
+		const CENTER_OFFSET: Vector3<f32> = Vector3{ x: 0.0, y: 5.0, z: 0.0};
 
 		let center  = center + CENTER_OFFSET;
 		let eye		= Point3::from_vec(center.to_vec() - (view_direction * DISTANCE));
@@ -40,7 +40,7 @@ impl Camera {
 		}
 	}
 
-	pub fn view_corners(&self) -> Vec<Vector3<f64>> {
+	pub fn view_corners(&self) -> Vec<Vector3<f32>> {
 		let view = self.view.transpose();
 		let view_origin = (view * self.view.w.clone()).truncate() * -1.0;
 		let right = view.x.truncate();
